requirements.txt
coloredlogs==15.0
bitstring==3.1.7
deepdiff==5.0.2
paho-mqtt==1.5.1
schedule==1.1.0